Add/remove empty lines
I added and removed empty lines, a.k.a. "whitespace". The browser
doesn't care how much whitespace we use in our HTML, CSS, and
JavaScript, but we can use whitespace to make our code easier to
read and edit. Empty lines in particular can help us to see what
blocks of code should be grouped together, or where a set of
related commands begins and ends.